requirements.txt
flask
flask-bcrypt
flask-sqlalchemy
python-dotenv
flask-login
flask-mail
datetime
os